Simplified Step
In order to simplify the code, the Step operation
has been split into 3 different effects. Step doesn't
regroup the 4 operations anymore:

- a Step
- a state getter
- a state setter
- a map operation

Mapping was already handled by Mapped, so there was no
need to add this effect to Step. To map a Step just
call the map operation on it!

The state getter and setter are now proper effects
decoupled from Step.